move downloadLog to export
